feat: add collection tag on report page
