+++ apps/web/src/modules/analyze/hooks/useCompareItems.ts
@@ -5,10 +5,12 @@ import { Level } from '@modules/analyze/constant';
 const useCompareItems = () => {
   const { verifiedItems } = useStatusContext();
 
-  const items = verifiedItems.map(({ label, level, shortCode }) => {
-    const name = level === Level.REPO ? getPathname(label) : label;
-    return { label, level, shortCode, name };
-  });
+  const items = verifiedItems.map(
+    ({ label, level, shortCode, collections }) => {
+      const name = level === Level.REPO ? getPathname(label) : label;
+      return { label, level, shortCode, name, collections };
+    }
+  );
 
   const ids = items.map((i) => i.shortCode);
